# SWAGGER_API_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: Lilium B2B E-commerce Backend API
  description: Complete API documentation for the Lilium B2B multi-vendor platform with cash-on-delivery system
  version: 1.0.0
  contact:
    name: Lilium API Support
    email: support@lilium.iq
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.lilium.iq/api
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /upload/single:
    post:
      tags: [Upload]
      summary: Upload single file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: File uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
  /upload/multiple:
    post:
      tags: [Upload]
      summary: Upload multiple files
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        200:
          description: Files uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  urls:
                    type: array
                    items:
                      type: string
...